+++ Data Structures/SegTreeRec.cpp
@@ -6,34 +6,34 @@ using namespace std;
 template <typename T> class SegTree { // Tudo 0-based
 
 public:
-  void build(size_t node, size_t L, size_t R, const vector<T> &V) {
+  void build(int node, int L, int R, const vector<T> &V) {
     if (L == R)
       Tree[node] = V[L];
     else {
-      size_t M = (L + R) / 2;
+      int M = (L + R) / 2;
       build(2 * node, L, M, V);
       build(2 * node + 1, M + 1, R, V);
       Tree[node] = op(Tree[2 * node], Tree[2 * node + 1]);
     }
   }
 
-  T query(size_t l, size_t r, size_t node, size_t L, size_t R) {
+  T query(int l, int r, int node, int L, int R) {
     if (l > R or r < L)
       return neutral;
     if (l <= L and R <= r)
       return Tree[node];
 
-    size_t M = (L + R) / 2;
+    int M = (L + R) / 2;
     return op(query(l, r, 2 * node, L, M), query(l, r, 2 * node + 1, M + 1, R));
   }
 
-  T query(size_t l, size_t r) { return query(l, r, 1, 0, sz - 1); }
+  T query(int l, int r) { return query(l, r, 1, 0, sz - 1); }
 
-  void update(size_t id, T val, size_t node, size_t L, size_t R) {
+  void update(int id, T val, int node, int L, int R) {
     if (L == R)
       Tree[node] = op(Tree[node], val);
     else {
-      size_t M = (L + R) / 2;
+      int M = (L + R) / 2;
 
       if (id <= M)
         update(id, val, 2 * node, L, M);
@@ -44,7 +44,7 @@ public:
     }
   }
 
-  void update(size_t id, T val) { update(id, val, 1, 0, sz - 1); }
+  void update(int id, T val) { update(id, val, 1, 0, sz - 1); }
 
   SegTree(const vector<T> &V, function<T(T, T)> op_, T neutral_)
       : neutral(neutral_), op(op_) {
@@ -54,13 +54,13 @@ public:
     build(1, 0, sz - 1, V);
   }
 
-  SegTree(size_t sz_, function<T(T, T)> op_, T neutral_)
+  SegTree(int sz_, function<T(T, T)> op_, T neutral_)
       : sz(sz_), neutral(neutral_), op(op_) {
     Tree = vector<T>(4 * sz, neutral);
   }
 
 private:
-  size_t sz;
+  int sz;
   vector<T> Tree;
   const T neutral;
   function<T(T, T)> op;
